Add files via upload
Day 1 of the Data Science Masters course taught by Krish Nayak Sir would focus on introducing the Python programming language and its use in data science. The following topics would be covered:

  *Introduction to Python and comparison with other programming languages
  *Python objects, numbers, booleans, and strings
  *Container objects and the mutability of objects

The course would be a mix of lectures, interactive discussions, and hands-on exercises to help participants build their understanding of the Python language and its features. By the end of Day 1, participants would have a solid foundation in the basics of Python programming and be ready to move on to more advanced topics in data science